requirements.txt
crewai
crewai_tools
load_dotenv
langchain-google-genai
duckduckgo-search
langchain-community
streamlit
asyncio  